requirements-docs.txt
Sphinx
docutils
sphinx-django-command
